Restructure project
The main change is the extraction of Piggybank update logic into the hook 'usePiggybank'. The rest is just regrouping functions and types into more "obvious" files. A few things were renamed for further clarity.
